Split out CARD ontology provider code from CARDTargetProvider to CARDTargetOntologyProvider
