+++ rcsb/utils/tests-targets/testCARDTargetOntologyProvider.py
@@ -1,7 +1,7 @@
 ##
-# File:    CARDTargetProviderTests.py
-# Author:  J. Westbrook
-# Date:    27-Nov-2020
+# File:    testCARDTargetOntologyProvider.py
+# Author:  D. Piehl
+# Date:    14-Mar-2023
 # Version: 0.001
 #
 # Update:
@@ -9,13 +9,13 @@
 #
 ##
 """
-Tests for utilities managing CARD target data.
+Tests for utilities managing CARD ontology data.
 
 """
 
 __docformat__ = "google en"
-__author__ = "John Westbrook"
-__email__ = "jwest@rcsb.rutgers.edu"
+__author__ = "Dennis Piehl"
+__email__ = "dennis.piehl@rcsb.org"
 __license__ = "Apache 2.0"
 
 import logging
@@ -25,7 +25,7 @@ import resource
 import time
 import unittest
 
-from rcsb.utils.targets.CARDTargetProvider import CARDTargetProvider
+from rcsb.utils.targets.CARDTargetOntologyProvider import CARDTargetOntologyProvider
 
 HERE = os.path.abspath(os.path.dirname(__file__))
 TOPDIR = os.path.dirname(os.path.dirname(HERE))
@@ -34,12 +34,9 @@ logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger()
 
 
-class CARDTargetProviderTests(unittest.TestCase):
+class CARDTargetOntologyProviderTests(unittest.TestCase):
     def setUp(self):
         self.__cachePath = os.path.join(HERE, "test-output", "CACHE")
-        #
-        self.__fastaPath = os.path.join(HERE, "test-output", "card-targets.fa")
-        self.__taxonPath = os.path.join(HERE, "test-output", "card-targets-taxon.tdd")
         #
         self.__startTime = time.time()
         logger.info("Starting %s at %s", self.id(), time.strftime("%Y %m %d %H:%M:%S", time.localtime()))
@@ -51,20 +48,23 @@ class CARDTargetProviderTests(unittest.TestCase):
         endTime = time.time()
         logger.info("Completed %s at %s (%.4f seconds)", self.id(), time.strftime("%Y %m %d %H:%M:%S", time.localtime()), endTime - self.__startTime)
 
-    def testFetchCARDTargets(self):
-        ctP = CARDTargetProvider(cachePath=self.__cachePath, useCache=False)
-        ok = ctP.testCache()
+    def testFetchCARDTargetOntology(self):
+        coP = CARDTargetOntologyProvider(cachePath=self.__cachePath, useCache=False)
+        ok = coP.testCache()
         self.assertTrue(ok)
-        ok = ctP.exportCardFasta(self.__fastaPath, self.__taxonPath)
-        self.assertTrue(ok)
+        aroIdL = ["ARO:3000096", "ARO:3000015"]
+        for aroId in aroIdL:
+            lineageL = coP.getLineage(aroId)
+            logger.info("Lineage list for aroId %s: %r", aroId, lineageL)
+            self.assertGreaterEqual(len(lineageL), 2)
 
 
-def fetchCARDTargets():
+def fetchCARDTargetOntology():
     suiteSelect = unittest.TestSuite()
-    suiteSelect.addTest(CARDTargetProviderTests("testFetchCARDTargets"))
+    suiteSelect.addTest(CARDTargetOntologyProviderTests("testFetchCARDTargetOntology"))
     return suiteSelect
 
 
 if __name__ == "__main__":
-    mySuite = fetchCARDTargets()
+    mySuite = fetchCARDTargetOntology()
     unittest.TextTestRunner(verbosity=2).run(mySuite)
